+++ .travis.yml
@@ -13,7 +13,7 @@ install:
   - pip install keras==$KERAS_VERSION
   - pip list
 script:
-  - nosetests --with-coverage --cover-package=dlhub_toolbox
+  - nosetests --with-coverage --cover-package=dlhub_sdk
   - cd examples
   - ./run_examples.sh
   - cd .. # So that coveralls works
