+++ dlhub_sdk/client.py
@@ -1,4 +1,4 @@
-from dlhub_toolbox.utils.schemas import validate_against_dlhub_schema
+from dlhub_sdk.utils.schemas import validate_against_dlhub_schema
 from tempfile import mkstemp
 import pandas as pd
 import requests
