+++ dlhub_sdk/tests/test_dlhub_client.py
@@ -1,5 +1,5 @@
-from dlhub_toolbox.models.servables.python import PythonStaticMethodModel
-from dlhub_toolbox.client import DLHubClient
+from dlhub_sdk.models.servables.python import PythonStaticMethodModel
+from dlhub_sdk.client import DLHubClient
 from unittest import TestCase, skip
 import pandas as pd
